requirements.txt
astral
psycopg2
Django
django-fsm
django-jinja
django-leaflet
djangorestframework
django-material
django-moderation
django-sequences
django-viewflow
django-waffle
django-widget-tweaks
celery
html5lib
rdflib-jsonld
newspaper3k
rdflib
plotly
PyYAML
Shapely
SPARQLWrapper
python-dateutil
Jinja2
django-filter
social-auth-app-django
ujson
